+++ package.json
@@ -25,7 +25,10 @@
   },
   "homepage": "https://github.com/LambdaSchool/node-api3-project#readme",
   "dependencies": {
+    "cors": "^2.8.5",
+    "dotenv": "^8.2.0",
     "express": "^4.17.1",
+    "helmet": "^4.1.1",
     "knex": "^0.21.2",
     "logger": "0.0.1",
     "sqlite3": "^5.0.0"
