+++ server.js
@@ -1,13 +1,61 @@
 const express = require('express');
-
+const cors = require('cors');
 const server = express();
+const userRouter = require("./users/userRouter");
+const postRouter = require("./posts/postRouter");
 
+server.use(express.json());
+server.use(cors());
+
+server.get('/api/users', logger, userRouter); 
+server.get('/api/posts', logger, postRouter);
 server.get('/', (req, res) => {
-  res.send(`<h2>Let's write some middleware!</h2>`);
+  res.send(`<h2>Hello! I am an API</h2>`);
 });
 
-//custom middleware
 
-function logger(req, res, next) {}
+// `logger` logs to the console the following information about each request: request method, request url, and a timestamp
+//      this middleware runs on every request made to the API
+
+function logger(req, res, next) {
+  console.log(
+    `[${new Date().toISOString()}]
+    ${req.method} to ${req.url} from ${req.get("Origin")}`
+
+  );
+  next();
+}
+
 
 module.exports = server;
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+// ##### Users
+
+// | field | data type        | metadata                                            |
+// | ----- | ---------------- | --------------------------------------------------- |
+// | id    | unsigned integer | primary key, auto-increments, generated by database |
+// | name  | string           | required, unique                                    |
+
+
+// ##### Posts
+
+// | field   | data type        | metadata                                            |
+// | ------- | ---------------- | --------------------------------------------------- |
+// | id      | unsigned integer | primary key, auto-increments, generated by database |
+// | text    | text             | required                                            |
+// | user_id | unsigned integer | required, must be the `id` of an existing `user`    |
